+++ tests/src/rules/default.js
@@ -22,8 +22,7 @@ ruleTester.run('default', rule, {
     // core modules always have a default
     test({ code: 'import crypto from "crypto";' }),
 
-    test({ code: 'import common from "./common";'
-         , settings: { 'import/ignore': ['common'] } }),
+    test({ code: 'import common from "./common";' }),
 
     // es7 export syntax
     test({ code: 'export bar from "./bar"'
@@ -76,17 +75,7 @@ ruleTester.run('default', rule, {
     }),
 
     test({
-      code: 'import crypto from "./common";',
-      settings: { 'import/ignore': ['foo'] },
-      errors: [{ message: 'No default export found in module.'
-               , type: 'ImportDefaultSpecifier'}]}),
-    test({
       code: 'import baz from "./named-exports";',
-      errors: [{ message: 'No default export found in module.'
-               , type: 'ImportDefaultSpecifier'}]}),
-
-    test({
-      code: 'import bar from "./common";',
       errors: [{ message: 'No default export found in module.'
                , type: 'ImportDefaultSpecifier'}]}),